Added check to survey shape file type
If the transects shapefiles are not of type line it will display a
warning message to the user and create a dummy shapefile with no
transects (well one transect with start point = end point). Additional
check in check.intersections means that this shapefile which will fails
to generate a gradient will not cause DSsim to crash.