Changed the way the density grid is constructed.
Turned out that jittering the points to ensure the coverage of the
study area in some cases introduced small discrepancies in x or y
values. This led to strange plotting behaviour. Rather than jittering
the points within the region  the construct get.surface.constant now
created a buffer around the region and retains all points within the
buffer. It does this using rgeos gBuffer