Allow the user to generate more animals for population generation in complex regions
Some complex and narrow study regions (e.g. fjords) will mean that a large number of the population generated within the density grid will in fact land outside the study region. Sometimes this situation is so severe that the requested population size cannot be achieved. This fix enables the user to increase the size of the population within the density grid to achieve the requested population size within the survey region.